+++ back/controller/Controllers.js
@@ -1,10 +1,9 @@
 const userController = require("./userController.js");
 const loginController = require("./loginController.js");
 const addProductController = require("./addProductController.js");
-const orderController = require("./orderController.js");
 const addOrderController = require("./addOrderController.js");
 const productController = require("./productController.js");
-const inventoryNINAController = require("./inventoryNINAController.js")
+const InventoryController = require("./InventoryController.js")
 const ordersController = require("./ordersController.js")
 const HomePageController =require("./HomePageController.js");
 const SearchBarController =require("./searchbarController.js");
@@ -13,10 +12,9 @@ module.exports = {
   userController,
   addProductController,
   loginController,
-  orderController,
   addOrderController,
   productController,
-  inventoryNINAController,
+  InventoryController,
   ordersController,
   HomePageController,
   SearchBarController,
